requirements.txt
pytest
pytest-html
flask~=3.1.2
pytest-playwright
playwright
python-dotenv
allure-pytest